Simplify gradle files
Try to remove some of the ODK team's specific workflow from
Collect. Many of the power user features the core ODK team
relies on aren't necessary for Collect. In the interest of
keeping the setup as simple as possible, this removes all
remote configs and removes signing parameters from the gradle
files.