Addresses issue #248 https://github.com/opendatakit/collect/issues/248
Adds new utility class 'ListViewUtils' with static helpers for checking and unchecking ListView items.

NOTE: Behavior of 'toggle' button is not actually a toggle. It does not flip the state of individual items.
Instead what is implemented is really two binary functions "Check All" and "Uncheck All".

A clean reimplementation would replace the label of the 'Toggle All' button with alternating 'Check All' and 'Uncheck All'
labels, dependent on the current state of checked items in the list.

This will be filed as a new bug.

Changes to be committed:
	modified:   collect_app/src/main/java/org/odk/collect/android/activities/FormDownloadList.java
	new file:   collect_app/src/main/java/org/odk/collect/android/utilities/ListViewUtils.java