Create a pull request template
I'd like to try using a template for pull requests so that we can get more information about each. The goal is to help make sure reviews are as meaningful as they can be without taking up too much time. Hopefully this will act as a helpful reminder that we should be verifying and explaining our changes when we submit PRs.