Disable back button on first question (#1761)
Previously we used

boolean firstQuestion = formController.stepToPreviousScreenEvent() == FormEntryController.EVENT_BEGINNING_OF_FORM;
to check if the current index is the first question and after that, we weren't in the original index. Now since we use

formController.isCurrentQuestionFirstInForm()
we check it and go back to the original index (the isCurrentQuestionFirstInForm() method does that for us) that's why we don't need to use

stepToNextScreenEvent
below anymore, we need stepToPreviousEvent() since we check if a repeat section exists in a lower index.