Fixes 1052
The fix now handles two scenarios:

1) Blank / Empty sheet

If a Geo Location field exists, then altitude / accuracy columns will be added to the sheet. The column titles will be the title of the
data field appended with "-accuracy" and "-altitude".

2) Existing Sheet

The "-altitude" and "-accuracy" columns will not automatically be added, by default existing functionality will be maintained.
However, if the columns are manually inserted into the sheet with the title of the Geo Location + "-accuracy" and "-altitude"
then those data items will be exported to the columns.