Replaces black icons with tint colored icons for Preference Screens
This was a fallback from replacing the icons with vectors. Vectors are working fine with lollipop and above devices but nothing is appearing in place of icons in pre-lollipop devices. I tried to create a custom view to enable the use of vectors but I went too far from the implementation of PreferencesActivity and it didn't feel like a good solution. So, until we have a better solution to replace the png's in preference-headers, we can continue to use the existing png's as drawables
''