Fix #943
This initial commit has removed the duplicate code around clicking on images in 5 classes: SignatureWidget, ImageWidget, DrawWidget, AlignedImageWidget, AnnotateWidget.

Each of those classes had the ability to click on an image / or button that launched a Drawing activity or a viewing activity.

To remove the duplication, I opted for creating an ImageClicker interface with two concrete classes; One for launching the viewer activity and one for launching the draw activity.

On construction, the 5 classes set an imageClickHandler with the handler that they require.

I opted to put the interface; ImageClicker and the two concrete implementations; ViewImageClickHandler and DrawImageClickHandler into the base class BaseImageWidget.

Also cleaned up QuestionWidget a little bit, it was setting the ImageView onClickListener and only performing an action if the object was of type BaseImageWidget. It made a lot more sense to relocate this to BaseImageWidget.