Fix #943
This commit tackles the duplication in the classes; AlignedImageWidget, AnnotateImage, and ImageWidget around the handlers for launching Activities for choosing and capturing images.

Each of the classes have very similar methods: chooseImage() and captureImage().

Following on from the design used for the imageClickHandler in the previous commit, i.e. declaring an interface and concrete implementation as an inner class inside of BaseImageWidget, I defined the interface ExternalImageCaptureHandler and the concrete class ImageButtonClickHandler inside of BaseImageWidget. The above classes then assign an instance of ExternalImageCaptureHandler to a handler variable at construction time.

The captureImage() method of the three classes now call captureImage() in ExternalImageCaptureHandler(), once they have created their unique intent objects.

The handlers for the button: R.id.choose_image now call through to the ExternalImageCaptureHandler.chooseImage()

I also added an abstract method to BaseImageWidget - addExtrasToIntent. The reason is to allow classes to put extra params into the intent that is used to launch the drawActivity - I spotted that AnnotateWidget needed to add an extra for screen orientation. By making it abstract in BaseImageWidget means that all future subclasses will get the opportunity to intercept the Intent and add extra parameters.