fix #943
Removed further duplication around the areas of launching an activity, a new method was created: LaunchActivityForResult in BaseImageWidget and called from within the click handlers. This further removed duplicate code in three places within BaseImageWidget.

Also manually tested and ran the automated tests to ensure that everything was working as it was before.