Fixes #620
The plan is to first make only one class dependent on httpclient. After analysis, it seems to make sense that WebUtils should be the one and only class that knows anything about
httpclient. Once this initial step is complete, the job should be simpler to swap out httpclient.

In this commit I have refactored DownloadFormsTask.downloadFile() by creating a new method in WebUtils to initialise the getDownloadInputStream(). This has completely broken the dependancy on httpclient in DownloadFormsTask. I have also changed the method signature to WebUtils.getXmlDocument to remove the need to pass a localContext, httpClient - this was done in a number of places and they were always the same objects, therefore the getXmlDocument is now responsible for creating those.

Made changes in DownloadFormListTask and WebUtilsTest as a result of changing getXmlDocument method.