Fixes #620
Continuing with the strategy of removing all the dependencies on httpclient from all classes but WebUtils, this Pull request has targeted InstanceServerUploader.

I relocated uploadOneSubmission into WebUtils and renamed it to uploadFile, this is a very large method but I concentrated on the extraction rather than refactoring the method (Something to be done when completely removing httpclient). As a result of the remove, it also required me to relocate ContentTypeMapping and the Outcome class into WebUtils. WebUtils also required access to getFileExtension, which resided in InstanceServerUploader, it made more sense to put that as a static method of the FileUtils class.