Fix #620
Moved getHttpContext() from Collect to WebUtils which allowed the removal of 3 x httpclient imports from Collect.

Removed all references to httpclient from ResponseMessageParser, rather than passing a HttpEntity class to the constructor, it takes a response string. The conversion of the HttpEntity is now done in WebUtils.

Added a small method to WebUtils to return the Mime type of plain text, this was so that OSMWidget no longer needed a dependency on WebUtils.

Had to also tweak WebUtilsTest to getHttpContext from WebUtils instead of Collect.