Refactored SMS job queue implementation and tests
The usage of the android priority job queue was removed and the evernote android job library was implemented. The SmsSenderJob was refactored and it's functionality was encapsulated in another class to improve it's testability since jobs aren't able to run in a test environment. Along with these changes the base for the new instance row design was created wich involves a custom progress bar and an extended CursorAdapter