Fix #620
Renamed (and relocated) DownloadResult to InputStreamResult - Seems to be a better name.

Did some small tentative refactoring on WebUtils.uploadFile() to help reduce the size of the method and also replace some of the variable names to “better” more understandable names. Such as “j” to fileIndex and “f” to file.

Extracted the routine which created a list of files in the parent directory - It wasn’t strictly necessary but it has helped to reduce the method size a little bit and was a quick win.