refactoed AppListActivity to use CursorAdapter instead of SimpleCursorAdapter
This was necessary because the InstanceUploaderList will need a custom CursorAdapter to facilitate the interactions that are necessary with the different views of each row. With this change a new CursorAdapter can be used to replace the SimpleCursorAdapter that was being used without breaking any of the other list based activities. Along with this the instance row layout was modified to include the new progress bar while still maintaining support the checkboxes