Created an Event Bus powered by RxRelay and RxJava
This event bus allows classes to subscribe for events by event class type. Once events are posted they are filtered and the observers that are listening for a specific class type receive them. RxRelay was used since having onComplete or onError within the stream could cause it to be damaged thus impacting the flow of events so without them the stream will remain intact no matter what.