Support guidance hint with settings (#2105)
Supported for <hint> content only. It is a description of the question that can be used to provide further guidance to enumerators. It is not meant to be shown in the client UI by default, but could be shown in a special view mode (e.g., for a training) or on printouts.

* created a layout and view binding mechanisms for all help/hint related constructs. This will allow for the hint and the hint's guidance to live within the same layout. Also added a key to preferences that will be used to store the guidance setting.

* added list preference for guidance setting

* fetched guidance hint from the associated FormEntryCaption based on the hint's text id and form type of guidance.

* added animations to collapse guidance setup.

* wrote integration tests to verify that guidance hint works.

* removed this visibility check on help related layouts because if the help text is already gone then it's parent layout won't have a height due to it having the WRAP_CONTENT property so the AnswerView will automatically be positioned beneath the previous layout.

* added the ability to store expanded/collapsed state of guidance hint. Performed several style fixes.

* integrated guidance hint with dark/light theme functionality

* Refactored hint guidance layout approach

The hint guidance was being attached by using CompoundDrwables of TextView. The issue with using this was that there wasn't alot of flexibility with the layout attributes that could be applied to the drawable eg. padding can be applied (which was causing the bug) but not margins so I removed the drawable and created a layout to host the guidance textview and an an imageview for the icon

* Help layout hidden for child widgets of DateTimeWidget

The TimeWidget and DateWidget's help layout was hidden so that only the parent widget's layout is displayed. Previous only the help text was being hidden thus these changes had to take place due to the introduction of the new layout structure for hint guidance.