Fixed issue taking place with guidanceVisibilityContentTest
The action of the tests were checking to see if a guidance hint would be displayed if the help text layout was clicked. A failure occurred because the animation that's being used to animate the expansion of the hint guidance layout would cause issues if an IdlingResource isn't used and it's generally recommended not to use animations during tests. To resolve the issue I simply tested the use case where the hint guidance is set to Yes, so it's just shown once the question has been rendered.