Added storage permission checks to all activities so when the permission is revoked it's handled gracefully.
Entry point activities were able to handle storage permissions but all other activities weren't so they crashed when the user manually revoked the permission. Now when a user revokes a permission the app shows a dialog about it and then quits. This approach is taken because the revoking of storage permissions is an edge case in a sense and if an user actually does it, it's easier for them to restart the app via an entry point than to have permission granted code in all activities.