Storage runtime permissions resolves #644 (#2119)
* moved view instantiation code above permission check since onResume is making calls to some elements and grabbing views from the layout won't be impacted by result of permission check.

* Added GrantPermissionRule so that the tests on API 23 and above would be granted all the necessary permissions before they run.

* Applied permission rules and checks to all Content Providers and Activities that can be launched from an intent.

* removed due to no other declarants of these permissions being present. they were added as a fix to permission dialogs that weren't showing up during debugging.

* refactored initialization statements that were pulled from onCreate of FormEntryActivity

* fixed a bug that occurred due to requestWindowFeature causing a crash

RequestWindowFeature should always be called   before making any view related calls such as SetContentView.

* Added storage permission checks to all activities so when the permission is revoked it's handled gracefully.

Entry point activities were able to handle storage permissions but all other activities weren't so they crashed when the user manually revoked the permission. Now when a user revokes a permission the app shows a dialog about it and then quits. This approach is taken because the revoking of storage permissions is an edge case in a sense and if an user actually does it, it's easier for them to restart the app via an entry point than to have permission granted code in all activities.