Issue #943 widget redundancy (#2161)
Removed the duplicate code around clicking on images in 5 classes: SignatureWidget, ImageWidget, DrawWidget, AlignedImageWidget, AnnotateWidget.

Each of those classes had the ability to click on an image / or button that launched a Drawing activity or a viewing activity.

To remove the duplication, I opted for creating an ImageClicker interface with two concrete classes; One for launching the viewer activity and one for launching the draw activity.

On construction, the 5 classes set an imageClickHandler with the handler that they require.

I opted to put the interface; ImageClicker and the two concrete implementations; ViewImageClickHandler and DrawImageClickHandler into the base class BaseImageWidget.

Also cleaned up QuestionWidget a little bit, it was setting the ImageView onClickListener and only performing an action if the object was of type BaseImageWidget. It made a lot more sense to relocate this to BaseImageWidget.

Tackles the duplication in the classes; AlignedImageWidget, AnnotateImage, and ImageWidget around the handlers for launching Activities for choosing and capturing images.

Each of the classes have very similar methods: chooseImage() and captureImage().

Following on from the design used for the imageClickHandler in the previous commit, i.e. declaring an interface and concrete implementation as an inner class inside of BaseImageWidget, I defined the interface ExternalImageCaptureHandler and the concrete class ImageButtonClickHandler inside of BaseImageWidget. The above classes then assign an instance of ExternalImageCaptureHandler to a handler variable at construction time.

The captureImage() method of the three classes now call captureImage() in ExternalImageCaptureHandler(), once they have created their unique intent objects.

The handlers for the button: R.id.choose_image now call through to the ExternalImageCaptureHandler.chooseImage()

I also added an abstract method to BaseImageWidget - addExtrasToIntent. The reason is to allow classes to put extra params into the intent that is used to launch the drawActivity - I spotted that AnnotateWidget needed to add an extra for screen orientation. By making it abstract in BaseImageWidget means that all future subclasses will get the opportunity to intercept the Intent and add extra parameters.

Removed further duplication around the areas of launching an activity, a new method was created: LaunchActivityForResult in BaseImageWidget and called from within the click handlers. This further removed duplicate code in three places within BaseImageWidget.