Draw ProgressBar after view has loaded and made animations optional.
Even though the ProgressBar gets instantiated there are times when not all child views are drawn so I utilized the post method of the main layout to trigger the drawing of the progress bar that way the bar is only drawn when all views are ready. Animations are now optional for cases like initial displaying of the current progress.