Updated the DTOs being used for managing the state and of the different operations that a submission goes through.
The Message class was updated to utilize the MessageStatus instead of a boolean to track whether a message was sent or is sending. This allowed a Mapper to be used to transfer the MessageResultStatus from the SentBroadcastReceiver to that of the Message in an easy manner by utilizing similar enums.