Updated the InstanceUploaderAdapter to display all states and interact with all the functionality of the SmsService and existing system.
The UI state of this list can now show progress and states of normal uploads along with sms submissions. There are also constructs in place to show more detail for sms submissions since they are sending updates from the background.
UI States were for MessageStatus so that when the SmsService receives a change and it's being propagated back to the UI via the event bus it can be rendered well. Each state has an associated icon and subtext that's generated. Along with that, the progress of the submission was binded to the UI as well.