SmsService integrated with the InstancesDao and the MultiPart workflow.
SmsService now performs auto deletion and marking submissions in the InstancesDao as completed.
All important events take place are now sent to the UI via the event bus so that users can know when a  sms submission doesn't exist and when all other events take place such as airplane mode and the sms layer. The submitForm functionality was also refactored since we aren't sending a single message at a time so now we run checks to see if all messages have been sent and if not then we trigger another send.