Go to MainMenuActivity after FormEntryActivity (#2234)
Added finish to onItemClicked listener code that triggers the FormEntryActivity to be opened. This was accidentally removed when the FormChooserList was being refactored to support storage permissions. The denied function of the listener is utilizing a method called finishActivities to clear the entire stack instead of finish to just do so for the current activity, so while removing the finish method this was one was removed as well.