Added the Foreground Receiver that intercepts the notification broadcast
This receiver intercepts the broadcast and aborts it so once the user is currently on the InstanceUploaderList no notifications are created.Along with these changes the adapter's onDestroy method was called to dispose of all the subscriptions that take place when a row subscribes to the event bus.

A function to set different background from the default was added so that the card views can standout.