Registered the File Provider that will grant external apps access to the app's file directory
This is required because starting in Android N files can't be exposed to external apps directly. They have to be provided via a ContentProvider which allows the app to retain all the permissions to the file but it temporarily grants them to the external app that's going to utilize it.