Added transport preference code changes.
A preferences changed listener was attached to the transport preference so that when it's modified the state of the SMS destination setting can react to that. So if the transport is set for internet the entire SMS setting should be disabled and greyed out. Also the SMS preferences  was removed from the ServerPreferences screen since it's no longer a protocol being shown by itself.