Forms that are encrypted won't be sent as SMS and the user will be made aware of that.
FormsDao was integrated with SmsService so that the encryption status of the form could be checked. This status was then sent to the UI layer ( instance list and notifications) via the event bus so that the user gets an updated message that encrypted forms can't be submitted. A new model was created to encapsulate the data that's needed to submit a form instead of passing each of them as parameters to the submitForm function.