Created a mocked version of FormsDao and refactored the parameters being passed to the SmsService
Since we are primarily testing that the SMS can't be sent and we are sure of the FormsDao's functionality it was mocked to simply return false when it's isFormEncrypted method is called within the function that's tested SMS form submission. The a fake/sample model for the submitForm function was also created to give the SmsService the test data it needed for execution.