Rank Widget (#2180)
Support for the rank question type introduced in JavaRosa 2.11. If the value of the question is blank, the widget is disabled and the user must double tap to enable it. Double tapping is used instead of single tapping so that users can swipe to the next question. 

The widget presents the list of items. Users can long press on one to slide it to a desired position. Removing the response sets a blank value and disables the widget until the user double taps.