Updated InstanceUploaderList with proper protocol checking.
SMS Submission was a protocol by itself in the past so when users pressed the upload button the current protocol would be checked to see if twas SMS so that network checks could be skipped. Now since SMS submissions are beneath other protocols the transport mechanism is checked to know how to proceed. This code will change when the Both functionality is introduced.