SmsSubmissionManager and all related tests were refactored.
This includes, changing the contract to reflect what's actually happening when submissions are being forgotten and not necessarily deleted. The implementation of the contract was tested directly
so that the submissions could be cleared afterwards without putting the forgetAll code within the contract itself. If necessary, in the future a new implementation of this contract can be done and this impl. can be switched out to test that the new manager conforms to the existing functionality.