Use uri_data key in intent appearance as intent data
Intent data is needed for actions such as SENDTO used to send email or SMS. This introduces support for a special uri_data key with as minimal of changes as possible.

If uri_data is found in the extras list, a placeholder URI is used as the data. This is important because the data is required to resolve the activity and the value of uri_data is not known before it runs through populateParameters. A future improvement would be to refactor populateParameters so it doesn't mutate the extras but instead returns values that are ready to be used either as data or extras.