Removed the enums that were transferring state throughout the entire SMS architecture
- SentMessageResult
- SmsSendResultStatus
- SmsStatus

These were all enums that were being used through the different classes to transfer state. They were replaced with the result codes that are being used by the SmsManager along with several custom result codes as well. This was decided since they didn't have much responsibility and they were adding a layer of complexity to the overall platform.

Along with this, the delivery reports code was removed as well since it's not being used.