Refactored SmsService handling of submission state
This involved renaming one of the custom results.

Fixing a bug that was causing a crash due to submission being deleted before notification is shown so now only when a submission is being resent is it deleted.

Correcting conditions that determine when failure messages are to be persisted as well.

Renamed job method since a single message isn't being added to the queue anymore but the job is sending a multi part message.