Corrected value being passed to SmsService's submitForm during auto-send
Setting autoSend to true here will ensure that when a finalized instance doesn't have a SMS tag it fails silently instead of triggering a notification. A notification is triggered when an user triggers an upload via the instance list. In at case because the user did that they need to be made aware of the fact that the form has no tags but in this case only forms with tags should spawn notifications.