InstanceUploaderAdapter now uses the transport medium of the submission to show it's status.
A submission that was done via SMS will only show that specific status when the transport is set to sms otherwise it will show a generalized status similar to Aggregate/Google implementations. Once no SmsSubmission model is found then the latest status will be shown which would be an internet upload.