Add "Select on send" option to support both transports functionality (#2409)
* Added strings to support both transports functionality

Values were added to support the new SMS upload button and also to power the "Select on send" transport option.

* Modified conditions in ServerPreferencesFragment for the new "Select on send" option

The conditions were only reacting to the internet and sms transports so they were modified to support "Both" where in SMS and Both make the SMS Submission settings enabled.

* Refactored button holder layout for the new button

Added the new SMS upload button that will be visible when Both transports is enabled. The layout was refactored to support all children of the button holder resizing based on the tallest child.

* Implemented Both transports functionality in the InstanceUploaderList

This involved setting up conditions to toggle the visibility of the SMS upload button, modifying the upload button text and also knowing which upload functionality to use depending on the button clicked.

* Upload button now gets the text set when the transport type changes.

* removed blank lines.

* Introduced an enum called Transport for submission transport types

This helps with reducing the cognitive load it takes to read the string comparisons to differentiate between transport types.

* Optimized Transport enum implementation

Converted the string comparisons in onUploadButtonsClicked and also did a String cast inside fromPreference to remove the redundant casts.