Added a condition in getSelectedAccount that checks to see if account was deleted.
Now when code paths call to getSelectedAccount a check is done to see if the account is deleted. An account can be deleted for a variety of reasons. Sometimes a user might be signing into an org temporarily or a domain expires so the account is removed. So putting this in place ensures that no exceptions are thrown if this occurs which would cause an unwanted crash.