Test to verify getSelectedAccount behaviour when Google account is deleted.
This test verifies the behaviour of the GoogleAccountsManager when an account is deleted from the device. This verifies that when the account is deleted that calling getSelectedAccount will cause the manager to return null which gets passed to other code paths that notify the user of the occurrence so that they can select another account.