Corrected extras variable name and removed constraints from Sms worker.
An error was being caused due to the wrong variable name being used to access the Worker's InputData so I referenced the correct one in the SmsSenderWorker. Also it was realized that constraints were being supplied to the WorkManager that stated that SMS can only be sent when network access is present. Removed that as well.