Fix crash when Google account is no longer available (#2483)
Users can remove a Google account at the Android level. This means pulling the username from SharedPreferences could result in a crash. Additionally, the account can be deleted with Google.

* Refactored GoogleAccountsManagerTest

- Each time an account name is persisted to SharedPreferences an account with that name is added to the mocked Credentials to power the new getSelectedAccount function.

- Since there's now a check within chooseAccount to determine if a account chooser or a settings info should be displayed, the arguments for the condition were mocked in favour of showing the account chooser dialog

- Used lambda expressions to optimize some of the mocks.

* Test to verify getSelectedAccount behaviour when Google account is deleted.

This test verifies the behaviour of the GoogleAccountsManager when an account is deleted from the device. This verifies that when the account is deleted that calling getSelectedAccount will cause the manager to return null which gets passed to other code paths that notify the user of the occurrence so that they can select another account.