Removed manual server polling since the WorkManager does this by default.
Currently the ServerPollingWorker would enqueue a one time job if the device doesn't have internet so that would execute when connectivity is restored. An engineer from the arch components team stated that this isn't necessary because when the worker misses it's execution window due to a missing constraint it runs again after the flex period so this was removed and the network constraint added to the PeriodicWorkRequest.