Upgrade robolectric to 4.0-beta-1, add Guava
Tests hang with both Robolectric 3.5.1 and 3.8. The explicit Guava dependency is because android-api-client imported a really old version that did not provide the methods Robolectric needs.

The Guava 27 changelog says 'Important: ListenableFuture is now available as a separate artifact (com.google.guava:listenablefuture:1.0) from the rest of Guava. As part of that change, Guava now has a real (needed at runtime) dependency on com.google.guava:failureaccess:1.0'. With 27, we were getting a conflict with failureaccess. Guava 27 is really new so better to just go to a version that works for now.