Use singleton database helper to avoid leaks
There were lots of resource leaks logged which I thought were cursors not being closed. It turns out it was actually SQLiteOpenHelper objects that were leaking. This is because d52ce0b stopped using the singleton pattern for the helper and instead started creating a new helper that was never closed with every query.