Show user a toast when they try to view server settings with corrupt preferences
Initial work for #2086. I wanted to drive this change out with a Robolectric Fragment
test but I ran into problems with the Robolectric version and PreferencesFragment. Will
have a deeper look into this as it'd be beneficial to able to do Fragment testing in the
project.

For now, I've extracted a use case object to handle loading preferences into the Fragment
and tested it.