Use query to get latest versions of forms (#2691)
* Use query to get latest versions of forms

Previously, the filtering was done after loading an unfiltered cursor. Doing the filtering involved several passes over the form list so was inefficient. It also involved using getFormsFromCursor and expecting the cursor to stay open which was inconsistent with other usage.

To use a query for the filtering, groupby is used. CursorLoaders don't provide a direct way to specify a groupby query so this introduces a new URI for the FormsProvider.

* Remove unused FormUtils class

* Remove unused parameter