Start StoragePermissionActivity if storage permission is missing
Up until now, we were showing a dialog (for non-entry point activities) which notifies the user that the app cannot function without storage permission and exits from the app. But that means we have to handle this permission check separately in rest of the places.
So the idea behind this is to have a common check for all activities and allow another chance to grant the permission.
We destroy the current activity as it might lead to some inconsistent state due to the missing permission and recreate it once the permission is granted