Check for storage permission in onPostCreate
We need to initialize the cursor or read files from the storage (disk/instance sync task) once the storage permission is checked. But we cannot do that if the check is in onCreate (Since, at this point nothing is initialized in the extending activity). Doing this check in onPostCreate solves this problem. Also, if the permission is not granted then the code which relies on that permission is never run (preventing crashes)