Call refreshCurrentView() only when the variable is required
In the last commit, I added the null check for getCurrentViewIfODKView() in onActivityResult() but that initializes the widget even when it is not needed (eg. Moving back from HierarchyActivity). So, this creates another instance of the widget which we are trying to solve.

The best approach I could come up with was to instantiate the widget if "currentView" is null. This was required when "Do not keep activities" was enabled on Motorola G5 Plus/Samsung Galaxy S4 and we try to click a photo using the default camera app