Store the real instance of widget too as a field
The real object must be used to mutate the state of the object. A spy is the copy of the real instance so any changes made to it does not get reflected back to the real object. In our case, the intention is to set a mocked version of PermissionUtils