Strip Dagger configuration back to essentials
This moves the Dagger setup back to the simplest setup (I could see) for
the app. The setup is now closer to that of the official generic Dagger
docs (https://google.github.io/dagger/users-guide).

There had been some good work to setup Dagger for Android specifically
but this is now a little out of date (current docs at
https://google.github.io/dagger/android) and also was only being used by
some classes created to setup a refactor that never came to be
(GeoActivity etc). As we find examples of places we need Android
specific Dagger setup we might want to pull some of that work back in.