PermissionUtils refactor (#2733)
* Adds a public parameterized constructor to PermissionUtils

* Create PermissionUtil object in QuestionWidget and add a getter

* Convert all static references when requesting runtime permissions

* Remove redundant static modifier from permission requesting methods

* Remove activity parameter from the methods

We already have access to an activity object passed to the constructor

* Make showAdditionalExplanation method non-static

* Extract common code for checking permission into a new method

* Extract common code for requesting permission check into new methods

The intention behind this refactor is to later use the common entry point for the Dexter library during unit tests. Then we can mock the requestPermission method and control the permission from within the unit test without having to do that separately for all methods

* Use the created method "requestPermissions" for requesting permissions

All of the runtime permissions are going through a single method. By controlling only this function, we can control the permission check during testing.

* Move static methods to the top of the class

* Create PermissionUtils only when QuestionWidget is created by Activity

This was failing the tests as in unit tests, we are using application context for building the widgets

* Shorten the name of utility methods