Clean up form 'breadcrumb' paths (#2762)
* Clean up some path-building code

Simpler, but now the form index (1) appears while in the repeat group
picker when it shouldn't. This will be addressed shortly.

* Only move path index back if on a question

Seems to be an old bug (since ~2012). This makes it so that the current
group is shown in the path as expected.

* Postpone the path multiplicity until the next group

Show it on the child instance, not the parent group.

* Simplify path building; deliminate multiplicity numbers

e.g. instead of `Friends (1)` it will show `Friends > 1` in the path.

* Number repeat items in front instead of parenthetically

e.g. instead of `Friends (1)` it will show `Friends > 1` or `1. Alice`
for repeat group items.

* Retain NonNull annotation for method wrapper

* Add back LRM character

To prevent RTL languages from moving the period to the front.

* PR fix: Handle groups at very beginning of form

The very first group in a form is auto-entered, so this steps back out.