Refactor QuestionWidget
Remove unused getter. Since the member variable is only being used
within the same class, so the getters can be dropped

Annotate methods of GuidanceHint as @NonNull to remove nullable warning.
Since it is an enum & the default value of GuidanceHint should be 'No'
instread of 'null', so it is safe to return 'No' and mark method as @NonNull