Moves media player related functionalities to Media Widget class (#2817)
* Adds new abstract class, MediaWidget

The purpose of this class would be to pull out all functionality related
to media player

* Extend widgets with Media Widget

Audio, grid, grid-multi and select widgets require media player. So the
plan is to extend MediaWidget in these classes and in later commits,
start moving the media related methods and fields into MediaWidget

* Override release() in MediaWidget

The super call is redundant as of now, but it is better than removing
the method from QuestionWidget class

* Move getter for media player to MediaWidget

* Move stopAudio() to MediaWidget

* Move playAudio() and playVideo()

Additionally, a check had to be added to ODKView to check if the widget
is of type MediaWidget before performing autoplay during loading of the
widget

* Move playAllPromptText() to MediaWidget

1. Adds autoplay parameter to playAllPromptText()
2. Removes separate methods for playing audio/video and uses the above
parameter to distinguish between type of autoplay

Since the name of method playAllPromptText() is generic, so it should
behave in a generic manner as well. Until now, it was only concerned
with autoplaying audio, while giving an impression that it is common
method for all kinds of prompts (audio/video)

* Separate init of textview and media views

* Init media and playColor in MediaWidget

* Implement AudioPlayListener in MediaWidget

* Move mediaPlayer init to MediaWidget

* Refactor QuestionWidget

Remove unused getter. Since the member variable is only being used
within the same class, so the getters can be dropped

Annotate methods of GuidanceHint as @NonNull to remove nullable warning.
Since it is an enum & the default value of GuidanceHint should be 'No'
instread of 'null', so it is safe to return 'No' and mark method as @NonNull

* Adds license header

* checkstyle fixes

* Improve log text message

* Fix naming convention from varNAME to varName