Configure spotbugs only for debug build
This is to prevent creation or execution of tasks for other builds to speed up build time. As recommended by gradle, either we can delay that by using Lazy task configuration. But since we only make use of debug type, so removing the config for other builds is the better option here