Add tests to verify launching intents in Widgets (#2734)
* Add ButtonWidgetTest class

* Add MockedPermissionUtils to control the runtime permission callbacks

I have overridden requestPermissions() which is the interface between the app and Dexter library. So the normal implementation uses the Dexter library and the mocked version can be configured to dynamically according to the needs of the test

* Store the real instance of widget too as a field

The real object must be used to mutate the state of the object. A spy is the copy of the real instance so any changes made to it does not get reflected back to the real object. In our case, the intention is to set a mocked version of PermissionUtils

* Use MockedPermissionUtils and add basic skeleton of button tests

Add methods to compare if the launched intent is similar to expected intent

* Do not display dialog when permissions denied in MockedPermissionUtils

* Add tests for AudioWidget

* Add tests for AlignedImageWidgetTest

* Add tests for AnnotateWidget

I had to add a try-catch block in AnnotateWidget as FileProvider.getUriForFile was throwing error in unit test. Hence, not checking for android.provider.MediaStore.EXTRA_OUTPUT

* Add tests for DrawWidget

* Add tests for SignatureWidget

* Add tests for ImageWidget

* Add tests for ImageWebViewWidget

* Add tests for GeoTraceWidget

Also adds a Shadow for PlayServicesUtil to override play services check

* Add tests for GeoShapeWidget

Also adds a Shadow for PlayServicesUtil to override play services check

* Add tests for GeoPointWidget

Also adds a Shadow for PlayServicesUtil to override play services check

* Add tests for VideoWidget

* Add tests for OSMWidget

I had to use try-catch block instead of proactively checking if intent is safe to launch. This is required to mock triggering of the intent for unit test

* Add tests for BarcodeWidget

* Add a todo for adding bearing widget tests

There is already an open issue: https://github.com/opendatakit/collect/issues/2057

* Fix PMD and checkstyle issues

* Initialize ActivityAvailability in ExStringWidget to prevent NPE

* Rename MockedPermissionUtils to FakePermissionUtils

* Use ids to click button and verify intents manually

* Remove tests from ButtonWidgetTest

* Fix broken tests after rebasing onto master

* Fix issues after rebasing